LayoutTests/imported/w3c/web-platform-tests/css/css-shadow-parts/serialization-expected.txt


FAIL Escape start parenthesis in ::part name. assert_equals: expected "::part(\\(foo)" but got "::part((foo)"
FAIL Escape start space in ::part name. assert_equals: expected "::part(bar\\ )" but got "::part(bar )"
PASS Collapse spaces in ::part names list. 
